Hold Cover Update!
Hold Covers can now manually set their origin, as well as enable if they should be behind the strum receptors (like holds), as well as having an alternative positioning method which positions them on hold ends instead of at the strums.

Also added note splashes and hold covers to the big "allStrumSprites" group in PlayState cuz of this layering change.